requirements.txt
fastapi
uvicorn
black
databases
sqlalchemy
pydantic
pydantic[dotenv]
asyncpg
psycopg2
starlette
passlib
bcrypt
python-multipart
PyJWT